requirements.txt
aiogram>=3.0
fastapi~=0.103.1
numpy >= 1.2
opencv-python-headless
pillow >= 9.0
python-multipart
pyyaml~=6.0.1
torch
uvicorn
aiohttp~=3.9.1
opencv-python
insightface
sqlalchemy